+++ pybamm/models/submodels/interface/base_interface.py
@@ -35,8 +35,6 @@ class BaseInterface(pybamm.BaseSubModel):
             self.reaction_name = ""  # empty reaction name for the main reaction
         elif reaction == "lead-acid oxygen":
             self.reaction_name = "oxygen "
-        elif reaction == "lithium-ion oxygen":
-            self.reaction_name = "oxygen "
         elif reaction == "SEI":
             self.reaction_name = "SEI "
         elif reaction == "lithium plating":
@@ -133,8 +131,6 @@ class BaseInterface(pybamm.BaseSubModel):
                 j0 = pybamm.Scalar(0)
             elif self.domain == "Positive":
                 j0 = param.p.prim.j0_Ox(c_e, T)
-        else:
-            j0 = pybamm.Scalar(0)
 
         return j0
 
@@ -149,8 +145,6 @@ class BaseInterface(pybamm.BaseSubModel):
             return self.phase_param.ne
         elif self.reaction == "lead-acid oxygen":
             return self.param.ne_Ox
-        else:
-            return pybamm.Scalar(0)
 
     def _get_average_total_interfacial_current_density(self, variables):
         """
@@ -191,7 +185,6 @@ class BaseInterface(pybamm.BaseSubModel):
         Domain = self.domain
         domain = Domain.lower()
         reaction_name = self.reaction_name
-        param = self.param
         j_scale = self.phase_param.j_scale
 
         if self.reaction == "lithium metal plating":
@@ -201,9 +194,6 @@ class BaseInterface(pybamm.BaseSubModel):
                 "Lithium metal plating current density [A.m-2]": j_scale * j,
             }
             return variables
-
-        i_typ = param.i_typ
-        L_x = param.L_x
 
         # Size average. For j variables that depend on particle size, see
         # "_get_standard_size_distribution_interfacial_current_variables"
@@ -231,8 +221,6 @@ class BaseInterface(pybamm.BaseSubModel):
     def _get_standard_total_interfacial_current_variables(self, j_tot_av):
         domain = self.domain.lower()
 
-        i_typ = self.param.i_typ
-        L_x = self.param.L_x
         j_scale = self.phase_param.j_scale
 
         if self.half_cell and self.domain == "Negative":
@@ -247,8 +235,6 @@ class BaseInterface(pybamm.BaseSubModel):
                 "current density": j_tot_av,
                 f"X-averaged {domain} electrode total interfacial "
                 "current density [A.m-2]": j_scale * j_tot_av,
-                f"X-averaged {domain} electrode total interfacial "
-                "current density per volume [A.m-3]": i_typ / L_x * j_tot_av,
             }
 
         return variables
@@ -257,7 +243,6 @@ class BaseInterface(pybamm.BaseSubModel):
         Domain = self.domain
         domain = Domain.lower()
         reaction_name = self.reaction_name
-        param = self.param
         j_scale = self.phase_param.j_scale
 
         if self.reaction == "lithium metal plating":
@@ -269,8 +254,6 @@ class BaseInterface(pybamm.BaseSubModel):
             }
             return variables
 
-        i_typ = param.i_typ
-        L_x = param.L_x
         # Size average. For j0 variables that depend on particle size, see
         # "_get_standard_size_distribution_exchange_current_variables"
         if j0.domain in [["negative particle size"], ["positive particle size"]]:
@@ -294,36 +277,35 @@ class BaseInterface(pybamm.BaseSubModel):
             "exchange current density [A.m-2]": j_scale * j0,
             f"X-averaged {domain} electrode {reaction_name}"
             "exchange current density [A.m-2]": j_scale * j0_av,
-            f"{Domain} electrode {reaction_name}"
-            "exchange current density per volume [A.m-3]": i_typ / L_x * j0,
-            f"X-averaged {domain} electrode {reaction_name}"
-            "exchange current density per volume [A.m-3]": i_typ / L_x * j0_av,
         }
 
         return variables
 
     def _get_standard_volumetric_current_density_variables(self, variables):
+        if self.half_cell and self.domain == "Negative":
+            return variables
+
         Domain = self.domain
         domain = Domain.lower()
         reaction_name = self.reaction_name
 
         a = variables[f"{Domain} electrode surface area to volume ratio"]
         a_av = variables[f"X-averaged {domain} electrode surface area to volume ratio"]
-        j = variables[f"{Domain} electrode{reaction_name} interfacial current density"]
+        j = variables[f"{Domain} electrode {reaction_name}interfacial current density"]
         j_av = variables[
-            f"X-averaged {domain} electrode{reaction_name} interfacial current density"
+            f"X-averaged {domain} electrode {reaction_name}interfacial current density"
         ]
         scale = self.param.i_typ / self.param.L_x
 
         variables.update(
             {
-                f"{Domain} electrode{reaction_name} volumetric "
+                f"{Domain} electrode {reaction_name}volumetric "
                 "interfacial current density": a * j,
-                f"X-averaged {domain} electrode{reaction_name} volumetric "
+                f"X-averaged {domain} electrode {reaction_name}volumetric "
                 "interfacial current density": a_av * j_av,
-                f"{Domain} electrode{reaction_name} volumetric "
+                f"{Domain} electrode {reaction_name}volumetric "
                 "interfacial current density [A.m-3]": scale * a * j,
-                f"X-averaged {domain} electrode{reaction_name} volumetric "
+                f"X-averaged {domain} electrode {reaction_name}volumetric "
                 "interfacial current density [A.m-3]": scale * a_av * j_av,
             }
         )
@@ -455,9 +437,7 @@ class BaseInterface(pybamm.BaseSubModel):
             j = pybamm.SecondaryBroadcast(j_xav, [f"{domain} electrode"])
 
         # j scale
-        i_typ = self.param.i_typ
-        L_x = self.param.L_x
-        j_scale = i_typ / (self.phase_param.a_typ * L_x)
+        j_scale = self.phase_param.j_scale
 
         variables = {
             f"{Domain} electrode {reaction_name}"
@@ -479,9 +459,7 @@ class BaseInterface(pybamm.BaseSubModel):
         Domain = self.domain
         domain = Domain.lower()
         reaction_name = self.reaction_name
-        i_typ = self.param.i_typ
-        L_x = self.param.L_x
-        j_scale = i_typ / (self.phase_param.a_typ * L_x)
+        j_scale = self.phase_param.j_scale
 
         # X-average or broadcast to electrode if necessary
         if j0.domains["secondary"] != [f"{domain} electrode"]:
@@ -499,12 +477,6 @@ class BaseInterface(pybamm.BaseSubModel):
             "exchange current density distribution [A.m-2]": j_scale * j0,
             f"X-averaged {domain} electrode {reaction_name}"
             "exchange current density distribution [A.m-2]": j_scale * j0_av,
-            f"{Domain} electrode {reaction_name}"
-            "exchange current density distribution"
-            + " per volume [A.m-3]": i_typ / L_x * j0,
-            f"X-averaged {domain} electrode {reaction_name}"
-            "exchange current density distribution"
-            + " per volume [A.m-3]": i_typ / L_x * j0_av,
         }
 
         return variables