+++ pybamm/parameters/lithium_ion_parameters.py
@@ -741,15 +741,6 @@ class DomainLithiumIonParameters(BaseParameters):
         """Rescaled dimensionless electrode electrical conductivity"""
         return self.sigma(T) * self.main_param.delta
 
-    def t_change(self, sto):
-        """
-        Dimensionless volume change for the electrode;
-        sto should be R-averaged
-        """
-        return pybamm.FunctionParameter(
-            f"{self.domain} electrode volume change", {"Particle stoichiometry": sto}
-        )
-
     def k_cr(self, T):
         """
         Dimensionless cracking rate for the electrode;
@@ -888,7 +879,8 @@ class ParticleLithiumIonParameters(BaseParameters):
         inputs = {
             "Electrolyte concentration [mol.m-3]": c_e,
             f"{self.domain} particle surface concentration [mol.m-3]": c_s_surf,
-            f"{self.domain} particle maximum concentration [mol.m-3]": self.c_max,
+            f"{self.phase_prefactor}Maximum {self.domain.lower()} particle "
+            "surface concentration [mol.m-3]": self.c_max,
             "Temperature [K]": T,
             f"{self.phase_prefactor}Maximum {self.domain.lower()} particle "
             "surface concentration [mol.m-3]": self.c_max,
@@ -929,7 +921,8 @@ class ParticleLithiumIonParameters(BaseParameters):
         """
         inputs = {
             f"{self.domain} particle stoichiometry": sto,
-            f"{self.domain} particle maximum concentration [mol.m-3]": self.c_max,
+            f"{self.phase_prefactor}Maximum {self.domain.lower()} particle "
+            "surface concentration [mol.m-3]": self.c_max,
         }
         return pybamm.FunctionParameter(
             f"{self.phase_prefactor}{self.domain} electrode "
@@ -1039,23 +1032,11 @@ class ParticleLithiumIonParameters(BaseParameters):
         Dimensionless volume change for the electrode;
         sto should be R-averaged
         """
-        inputs = {
-            f"{self.domain} particle stoichiometry": sto,
-            f"{self.domain} particle maximum concentration [mol.m-3]": self.c_max,
-        }
         return pybamm.FunctionParameter(
-            f"{self.domain} electrode volume change", inputs
-        )
-
-    def k_cr(self, T):
-        """
-        Dimensionless cracking rate for the electrode;
-        """
-        T_dim = self.main_param.Delta_T * T + self.main_param.T_ref
-        delta_k_cr = self.E ** self.m_cr * self.l_cr_0 ** (self.m_cr / 2 - 1)
-        return (
-            pybamm.FunctionParameter(
-                f"{self.domain} electrode cracking rate", {"Temperature [K]": T_dim}
-            )
-            * delta_k_cr
+            f"{self.domain} electrode volume change",
+            {
+                "Particle stoichiometry": sto,
+                f"{self.phase_prefactor}Maximum {self.domain.lower()} particle "
+                "surface concentration [mol.m-3]": self.c_max,
+            },
         )